% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_function.R
\name{random.Date}
\alias{random.Date}
\alias{age.cal}
\title{Date Related Functions}
\usage{
random.Date(
  from,
  to = Sys.Date(),
  size,
  replace = FALSE,
  prob = NULL,
  by = "day"
)

age.cal(from, to = Sys.Date(), by = "year")
}
\arguments{
\item{from}{starting date. Required.}

\item{to}{end date.Optional. Optional. Default is today (`sys.Date()`).}

\item{size}{a non-negative integer giving the number of items to choose.}

\item{replace}{should sampling be with replacement?}

\item{prob}{a vector of probability weights for obtaining the elements of the vector being sampled.}

\item{by}{increment of the sequence.}
}
\value{
A vector of class "Date".
}
\description{
\code{random.Date()} is a function that randomly creates a vector of dates at a specified sample size between a specified date and a date.
\code{age.cal()} is a function that calculates the number of years (age by default), months, and days from a specified date to a specified date.
}
\seealso{
\code{\link[base]{Date}}, \code{\link[base]{seq.Date}}, \code{\link[base]{sample}}
}
