% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_not_intger_value.R
\name{find.not.integer.value}
\alias{find.not.integer.value}
\title{Find Where Values in the Vector that non-integer}
\usage{
find.not.integer.value(x, where = c("number", "logical"))
}
\arguments{
\item{x}{vector or data frame to examine.}

\item{where}{Choose whether to indicate where there are values that cannot be converted to numbers, either as numbers or logical types.}
}
\description{
This function is used to find a non-integer value in a vector
The input value can be of any type, but it must be a vector of numbers only.
If a string or other value is entered, a warning message will be displayed and NA will be returned.
}
\seealso{
If a data frame of more than two dimensions is entered, returns the column numbers where the columns contain values that cannot be converted to numbers.
}
