% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head_tail.R
\name{head_tail}
\alias{head_tail}
\title{Return Parts of an Data Frame}
\usage{
head_tail(data, n = 3L, col_n = 3L)
}
\arguments{
\item{data}{an data frame}

\item{n}{an integer. Specify the number of rows you want to extract. The data is extracted from the beginning and end of the specified number of lines, so the data frame is the specified number of lines x 2 + ellipsis.}

\item{col_n}{an integer. Specify the number of rows you want to extract. The data is extracted from the beginning and end of the specified number of columns, so the data frame is the specified number of columns x 2 + ellipsis.}
}
\description{
Only display the specified number of rows and columns of the data frame are extracted, otherwise "..." and abbreviations are used to denote the rest.
}
\details{
If the specified number of rows and columns is greater than the original data frame, the data is display as it is.
}
\examples{
head_tail(iris)

}
