% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list2dataframe.R
\name{list2data.frame_cbind}
\alias{list2data.frame_cbind}
\alias{list2data.frame_rbind}
\title{List Convert to dataframe}
\usage{
list2data.frame_cbind(list)

list2data.frame_rbind(list)
}
\arguments{
\item{list}{list to be converted to a data frame}
}
\description{
Function to convert a list into a dataframe.
For list of different lengths, the data frame is constructed according to the longest list, and for short lists, the missing places are filled with NA according to the long list.
\code{list2data.frame_cbind} makes each element of the list a column.
\code{list2data.frame_rbind} makes each element of the list a row.
}
