% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LittleMCAR_test.R
\name{LittleMCAR_test}
\alias{LittleMCAR_test}
\title{Little's missing completely at random (MCAR) test}
\usage{
LittleMCAR_test(data, ...)
}
\arguments{
\item{data}{A data frame}

\item{...}{
  Arguments passed on to \code{\link[=mlest2]{mlest2}}
  \describe{
    \item{\code{max_iterlim}}{Numeric. Upper limit of the number of iterations to avoid infinite loops.}
    \item{\code{iterlim}}{a positive integer specifying the maximum number of
    iterations to be performed before the program is terminated.}
  }}
}
\value{
A list with class \code{"htest"} containing the following components:
\item{statistic}{Chi-square value.}
\item{parameter}{Degrees of freedom used for chi-square.}
\item{p.value}{the p-value for the test.}
\item{method}{a character string indicating the type of test performed.}
\item{data.name}{a character string giving the name of the data.}
\item{missing.patterns}{Number of missing data patterns.}
\item{amount.missing}{Amount and percent of mssing data.}
\item{data}{The data, organized my missing data patterns.}
\item{stop.code}{The stop code returned by \code{\link[stats]{nlm}}.}
\item{iterations}{The number of iterations used by \code{\link[stats]{nlm}}.}
}
\description{
Use Little's (1988) test statistic to assess if data is missing completely at random (MCAR). The null hypothesis in this test is that the data is MCAR, and the test statistic is a chi-squared value.
In the \code{mlest()} function of the \code{mvnmle} package, the correct statistics could not be calculated because the solution of ML estimation may not converge with the default value.
Using a modified version of the \code{mlest2()} function, an algorithm is used to perform ML estimation until the solution converges.
}
\details{
If \code{stop.code} is \code{4}, the solution of maximum likelihood estimation has not converged at the upper limit of the number of calculations set by \code{max_iterlim};
 change \code{max_iterlim} and recalculate with an increased upper limit.
}
\note{
Code is adapted from \code{mcar_test} in \code{naniar} package.
}
\examples{
LittleMCAR_test(airquality)

}
\seealso{
\code{\link[infun]{mlest2}}
}
