% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode_.R
\name{mode_}
\alias{mode_}
\title{Calculate mode}
\usage{
mode_(...)
}
\arguments{
\item{...}{Accepts vectors and data frames.}
}
\description{
Given a vector or a data frame, it calculates the mode and frequency.
}
\details{
If multiple columns of data frames are given, the most frequent combination of combinations and frequencies is computed.
Large data frames cannot be calculated properly.
}
