% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode_.R
\name{mode_data.frame}
\alias{mode_data.frame}
\title{Calculate mode of data frame}
\usage{
mode_data.frame(data)
}
\arguments{
\item{data}{a data frame.}
}
\description{
Calculate the mode frequency for each column of the data frame.
}
\details{
The result is in the form of a data frame that returns answers in the form of column name, mode, and frequency.
More than one answer may be returned for a column as the mode may not be uniquely obtained.
}
\examples{
mode_data.frame(iris)
}
