% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_omit_select.R
\name{na.omit_select}
\alias{na.omit_select}
\title{Handling of NA in selected columns in a data frame}
\usage{
na.omit_select(.data, ..., .retrieve = TRUE)
}
\arguments{
\item{.data}{data frame}

\item{...}{One or more unquoted expressions separated by commas. Variable names can be used as if they were positions in the data frame, so expressions like x:y can be used to select a range of variables.}

\item{.retrieve}{logical If TRUE, a data frame is returned excluding rows containing NA in the selected columns; if FALSE, a data frame is returned with only rows containing NA in the selected columns.}
}
\description{
If NA is present in a selected column in the data frame, returns a data frame with the rows containing NA in that column removed if the default is the case.
If \code{.retrieve = FALSE} is specified, only rows with NA in the chosen column are returned.
Multiple columns may be specified as the columns to be selected.
}
