% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_moji.R
\name{rand_moji}
\alias{rand_moji}
\title{Create a random Japanese (Kanji or Hiragana) string}
\usage{
rand_moji(length, size, replace = TRUE, moji = c("kanji", "hiragana"))
}
\arguments{
\item{length}{a positive number, the length of the output string.}

\item{size}{a positive number, the number of strings to output.}

\item{replace}{should sampling be with replacement?}

\item{moji}{Select the string to output. Select Kanji or Hiragana.}
}
\value{
a vector.
}
\description{
Function to create a random Japanese (Kanji or Hiragana) string.
}
\details{
It is a random string, so it does not reflect the normal rules of Japanese.
In the case of hiragana, characters that do not normally appear at the beginning of a string, such as Sutegana and "n", will also appear at the beginning.

Only the range of regular kanji is supported.
}
\examples{
rand_moji(length = 3, size = 3, moji = "kanji")
rand_moji(length = 3, size = 3, moji = "hiragana")
}
