% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readme.R
\name{readme}
\alias{readme}
\title{Access The Package README in a Browser}
\usage{
readme(package)
}
\arguments{
\item{package}{The name of the package whose README you want to refer to. String. You can specify only one package, and it must be installed.}
}
\description{
Access The Package README in a Browser.
With the package installed, access the README of the installed package from CRAN or GitHub with a browser.
}
\details{
We refer to the package's Description to determine whether it was installed from CRAN, GitHub, or otherwise.

If the package was installed from CRAN, it accesses the CRAN package web page with the README; if there is no README, an empty web page is displayed.

If the package was installed from GitHub, the web page of package on the GitHubis accessed.
}
