% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repcale_match.R
\name{replace_match}
\alias{replace_match}
\title{Replace the value that exactly matches the pattern with a replacement}
\usage{
replace_match(x, pattern, replacement, nomatch)
}
\arguments{
\item{x}{Vector of values to be converted.}

\item{pattern}{The value of the pattern to be converted or a vector containing it.
If given in a vector of two or more lengths, it is determined that it should be replaced by a value equal to the position of each of the values given in the replacement argument.}

\item{replacement}{The value of the replacement to be converted or a vector containing it.
If given in a vector of two or more lengths, it is determined that it should be replaced by a value equal to the position of each of the values given in the pattern argument.}

\item{nomatch}{Specifies the value to return if there is no matching pattern and replacement value. If no value is specified, the original x value is returned.}
}
\description{
Function to replace a value that exactly matches a pattern with a replacement.
Given a vector of equal length for pattern and replacement, the first value of pattern is interpreted as replacing the first value of replacement.
See the argument descriptions for details.
}
\examples{
values <- c("HOKKAIDOU", "hokkaidou", "TOUHOKU", "touhoku")
pattern <- c("hokkaidou", "touhoku")
replacement <- c("HOKKAIDOU", "TOUHOKU")
replace_match(values, pattern, replacement)
replace_match(values, pattern, replacement, nomatch = NA)
}
