% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_any.R
\name{round_any}
\alias{round_any}
\alias{round_any_ceiling}
\alias{round_any_floor}
\title{Rounding to an Arbitrary Value}
\usage{
round_any(x, by, type = c("ceiling", "floor"), origin = 0)

round_any_ceiling(x, by)

round_any_floor(x, by)
}
\arguments{
\item{x}{a numeric vector.}

\item{by}{Specify a positive value for the isoperimetric difference of the sequence of isoperimetric numbers for which you want to obtain an approximation.}

\item{type}{Specifies whether the value to be rounded is below or above the nearest value.}

\item{origin}{Specifies the first term of an identity sequence to be compared. The default value is 0.}
}
\description{
Outputs a sequence of numbers in which a given vector is close to a sequence of numbers in which an arbitrary value (positive) is specified and that value is the equal difference.
You can choose how to apply the close values by specifying \code{ceiling} or \code{floor} as the \code{type} argument to make it behave like rounding up or down.
If the value matches an arbitrary isoperimetric sequence, the value will be output as is.
\code{round_any_ceiling} is a simplified version of \code{round_any}, which outputs the result with the argument of \code{type} fixed to \code{ceiling} and \code{origin} fixed to \code{0}.
\code{round_any_floor} is a simplified version of \code{round_any}, where the \code{type} argument is fixed to \code{floor} and the \code{origin} is fixed to \code{0}.
\code{round_any_*} is faster than \code{round_any} in most cases, because the internal processing is done as a vector.
However, in rare cases, \code{round_any_*} may not be possible to obtain accurate values because of R's internal floating point arithmetic. \code{round_any} creates a sequence of numbers and compares them, so it gives accurate rounding results.
}
