% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/same_value_col.R
\name{same.value.col}
\alias{same.value.col}
\title{Find Columns Consisting of the Same Value in a Data Frame}
\usage{
same.value.col(data, col.name = c("number", "name"))
}
\arguments{
\item{data}{Data frame to check if there is a column consisting of the same value.}

\item{col.name}{Select whether the columns are represented by column names or by column numbers. The default value is the column number.}
}
\value{
a list.
}
\description{
This function is used to find a column consisting of the same value in a data frame.
The answer will be returned in a list format, and those that have the same value in the column number or column name will be output.
}
