% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_gtsummary.R
\name{save_gtsummary}
\alias{save_gtsummary}
\title{Outputs the Table Created by the gtsummary package in Various File Formats.}
\usage{
save_gtsummary(
  x,
  path,
  values = NULL,
  pr_section = NULL,
  zoom = 3,
  expand = 10,
  webshot = "webshot"
)
}
\arguments{
\item{x}{gtsummary table object.}

\item{path}{PowerPoint or Word or Image file to be created. It should end with filename extensions. Supported filename extensions: .pptx, .docx, .png, .pdf, .jpg.}

\item{values}{a list (possibly named), each element is a flextable object. If named objects, names are used as slide titles.}

\item{pr_section}{a \code{\link[officer]{prop_section}} object that can be used to define page layout such as orientation, width and height.}

\item{zoom}{parameters used by \code{webshot} function.}

\item{expand}{parameters used by \code{webshot} function.}

\item{webshot}{webshot package as a scalar character, one of "webshot" or "webshot2".}
}
\description{
This function is used to output the table created by the gtsummary package in PowerPoint or Word, or as an image file.
It just wraps \code{gtsummary}'s \code{\link[gtsummary]{as_flex_table}} and \code{flextalbe}'s \code{save_as*} functions.
}
\details{
The \code{\link[gtsummary]{as_flex_table}} argument of \code{gtsummary} is the default, so if you want to fine-tune it, you can do it yourself.
}
