% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_dataframe.R
\name{scale.data.frame}
\alias{scale.data.frame}
\title{Scaling and Centering of Data Frame Objects}
\usage{
\method{scale}{data.frame}(x, center = TRUE, scale = TRUE)
}
\arguments{
\item{x}{a data frame object.}

\item{center}{either a logical value or numeric-alike vector of length equal to the number of columns of data frame, where ‘numeric-alike’ means that \code{\link[base]{as.numeric}}(.) will be applied successfully if \code{\link[base]{is.numeric}}(.) is not true.}

\item{scale}{either a logical value or a numeric-alike vector of length equal to the number of columns of data frame.}
}
\value{
a data frame object.
}
\description{
\code{scale.data.frame} is generic function whose default method centers and/or scales the columns of a numeric in data frame.
The non-numeric values in the data frame will remain unchanged.

In short, it is a generic function of \code{scale}.
}
\seealso{
\code{\link[base]{scale}}
}
