% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_extract_sandwich.R
\name{str_extract_sandwich}
\alias{str_extract_sandwich}
\title{Extracts strings sandwiched between specified strings}
\usage{
str_extract_sandwich(string, start_pattern, end_pattern, perl = FALSE)
}
\arguments{
\item{string}{Input vector. Either a character vector, or something coercible to one.}

\item{start_pattern}{Specifies the first character of the string to be extracted. It must be different from the last character, \code{end_pattern}.}

\item{end_pattern}{Specifies the last character of the string to be extracted. It must be different from the first character, \code{start_pattern}.}

\item{perl}{logical. Should Perl-compatible regexps be used?}
}
\description{
Extracts strings sandwiched between specified strings.
}
\details{
The specified string must be a single character or three characters including the escape, and the first and last characters of the string must be different.
}
