% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_interchange_col.R
\name{subset_interchange_col}
\alias{subset_interchange_col}
\title{Look for any two columns of data frames that are interchangeable}
\usage{
subset_interchange_col(data, x, y, out.put = c("dataframe", "num"))
}
\arguments{
\item{data}{frame to examine.}

\item{x}{The column name to look up. The column name should be enclosed in double quotation marks.}

\item{y}{The column name to look up. The column name should be enclosed in double quotation marks.}

\item{out.put}{Selects whether to return the resulting output as a dataframe or by row name. The default is to return in a data frame.}
}
\value{
Returns either the extracted data frame or a vector of row counts.
}
\description{
For any two columns specified in the data frame (say column A and B), if the combination of column A and B is the same even if they are swapped, it will return it as a data frame or a row number.
For example, if column A has "TOM" and "BOB", and the same respective row in column B has "BOB" and "TOM "BOB" and "TOM" in the same row of column B, the row will be extracted as interchangeable.
Also, when there is a row with the same value in column A and B, it is also determined to be interchangeable and extracted.
}
