% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableone_overall_rename.R
\name{tableone.overall.rename}
\alias{tableone.overall.rename}
\title{Change the Text of "Overall" in a Table created with tableone}
\usage{
tableone.overall.rename(x, rename.str)
}
\arguments{
\item{x}{Object returned by CreateTableOne function contain the "Overall" character in the item name of a table.}

\item{rename.str}{The string that want to change from "Overall".}
}
\value{
An object of class \code{TableOne}, which is a list of three objects.
}
\description{
This function is used to change the "Overall" character in the item name of a table created with \code{\link[tableone]{CreateTableOne}} to any character.
}
\seealso{
\code{\link[tableone]{CreateTableOne}}, \code{\link[tableone]{print.TableOne}}, \code{\link[tableone]{summary.TableOne}}
}
