% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableone_rename_headline.R
\name{tableone.rename.overall}
\alias{tableone.rename.overall}
\alias{tableone.rename.headline}
\title{Change the Text of headline in a Table created with tableone}
\usage{
tableone.rename.overall(x, rename.str)

tableone.rename.headline(x, rename.headline)
}
\arguments{
\item{x}{Object returned by CreateTableOne function contain name of the heading to be changed.}

\item{rename.str}{The string that want to change from "Overall".}

\item{rename.headline}{Enter the name of the heading to be changed in formula form.
If you want to change more than one heading, combine the formulas into a list format.
Names that contain hyphens will be evaluated as negative in the formula, so they must be enclosed in quotation marks.}
}
\value{
An object of class \code{TableOne}, which is a list of three objects.
}
\description{
These functions are used to change the headline character in the item name of a table created with \code{\link[tableone]{CreateTableOne}} to any character.
\code{tableone.rename.overall} is a function that changes only the Overall of the table heading to an arbitrary character.
\code{tableone.rename.headline} is a function that change any heading (including Overall) to any character by setting the table heading as an formula before and after the change.
}
\seealso{
\code{\link[tableone]{CreateTableOne}}, \code{\link[tableone]{print.TableOne}}, \code{\link[tableone]{summary.TableOne}}
}
