% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_times.R
\name{to_times}
\alias{to_times}
\title{To Times Objects from POSIX or Chron Objects}
\usage{
to_times(x, tz = "")
}
\arguments{
\item{x}{\code{POSIXct} or \code{PSIXlt} objcet or \code{chron} object to be converted to \code{times} object.}

\item{tz}{a character string specifying the time zone to be used for
    the conversion.  System-specific (see \code{\link[base]{as.POSIXlt}}), but
    \code{""} is the current time zone, and \code{"GMT"} is UTC.
    Invalid values are most commonly treated as UTC, on some platforms with
    a warning.}
}
\description{
Create a \code{times} object in the \code{chron} package by taking only the elements of time from a \code{POSIXlt} or \code{POSIXct} object or \code{chron} object in \code{chron} package.
}
\details{
If the \code{POSIXct} object differs from the system time zone, the time zone must also be specified within \code{to_times}.
}
\examples{
x <- as.POSIXlt(Sys.time())
to_times(x)
}
