% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_.R
\name{var_}
\alias{var_}
\title{Interval Estimation of Population Variance and Hypothesis Testing}
\usage{
var_(
  x,
  alternative = c("two.sided", "less", "greater"),
  pv = 0,
  conf.level = 0.95
)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.}

\item{pv}{a number indicating the true value of the population variance.}

\item{conf.level}{confidence level of the interval.}
}
\description{
Compute an interval estimate of the population variance of x and a hypothesis test using the given population variance.
}
\details{
The sample variance of the estimate is the unbiased variance computed with \code{stats::var}. It also calculates the population variance assuming the given value is the population.
}
