% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_code.R
\name{dummy_code}
\alias{dummy_code}
\title{Create dummy coded variables}
\usage{
dummy_code(x, split, prefix = NULL, suffix = NULL, ...)
}
\arguments{
\item{x}{a vector or one column of data frame to be transformed into dummy codes}

\item{split}{a character (or object which can be coerced to such) containing regular expression to use for splitting. If empty matches occur, in particular if split has length 0, x is split into single characters.}

\item{prefix}{a String to be prefix to the column name in the dummy code data frame}

\item{suffix}{a String to be suffix to the column name in the dummy code data frame}

\item{...}{Other arguments to carry over to \code{\link[base]{strsplit}}.}
}
\value{
a data.frame of dummy coded variables.
}
\description{
Given a variable x with n distinct values, create n new dummy coded variables coded 0/1 for presence (1) or absence (0) of each variable.
This function can be used to create a dummy code by splitting a single value into multiple values separated by commas or other delimiters by specifying any delimiter character.
}
\examples{
df_sample <- data.frame(sample = c("a,b", "b", "c", "c,a", "a,b,c"))
(df_dummy <- dummy_code(df_sample$sample, split = ","))
new_df <- cbind(df_sample, df_dummy)
new_df
}
