% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_not_Date_value.R
\name{find.not.Date.value}
\alias{find.not.Date.value}
\title{Find Where Values in the Vector that cannot be Converted to Date}
\usage{
find.not.Date.value(x, where = c("number", "logical"))
}
\arguments{
\item{x}{vector to examine.}

\item{where}{Choose whether to indicate where there are values that cannot be converted to numbers, either as numbers or logical types.}
}
\description{
This function is used to find the where in the vector there are values that cannot be converted to \code{\link[base]{Date}} using \code{\link[lubridate]{as_date}} in \code{lubridate}.
The result will be the location of the value that cannot be converted to a number in the vector by default.
}
\note{
There is a slight difference between the values that can be converted to Date by \code{lubridate}'s \code{as_date} and those that can be converted by \code{base}'s \code{as.Date}.
\code{as_date} converts even relatively fuzzy forms if they can be changed to a \code{Date} class, while \code{as.Date} operates relatively more strictly.
}
