% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hosmer_test.R
\name{hosmer_test}
\alias{hosmer_test}
\title{Hosmer-Lemeshow Goodness of Fit (GOF) Test}
\usage{
hosmer_test(model, g = 10)
}
\arguments{
\item{model}{a \code{glm} -object with binomial-family.}

\item{g}{numeric, the number for how many subgroups the data should be divided into.}
}
\description{
Hosmer-Lemeshow Goodness of Fit (GOF) Test is to check model quality of logistic regression models.
Note that this function has a unique way of dividing subgroups. See details.
}
\details{
The Hosmer-Lemeshow Goodness of Fit (GOF) Test is a method for obtaining statistics by dividing observed and expected values into several arbitrary subgroups.
The method of dividing the observed and expected values into subgroups is generally based on the quantile of the expected value, for example, by taking a decile of the expected value.
This method is used in the \code{hoslem.test()} function of the \code{resouceselection} package and the \code{performance_hosmer()} function of the \code{performance} package.
However, there are several variations on how to divide the subgroups, and this function uses a method in which the expected values are ordered from smallest to largest so that each subgroup has the same number of samples.
}
\examples{
data("Titanic")
df <- data.frame(Titanic)
df <- data.frame(Class = rep(df$Class, df$Freq),
                 Sex = rep(df$Sex, df$Freq),
                 Age = rep(df$Age, df$Freq),
                 Survived = rep(df$Survived, df$Freq))
model <- glm(Survived ~ . ,data = df, family = binomial())
hosmer_test(model)
}
\references{
David W. Hosmer, Stanley Lemesbow (1980). Goodness of fit tests for the multiple logistic regression model, Communications in Statistics - Theory and Methods, 9:10, 1043-1069, \doi{10.1080/03610928008827941}
}
