% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavel_vertical.R
\name{label_vertical}
\alias{label_vertical}
\alias{vertical_list}
\title{Labels to Vertical Writing System}
\usage{
label_vertical(replace_list = vertical_list())

vertical_list()
}
\arguments{
\item{replace_list}{Specifies what to replace the string with, either in expression form or in list form containing expression forms. If no replacements are needed, specify NULL.}
}
\description{
\code{label_vetical} is function to convert the axis labels of a ggplot2 format graph to a vertical writing system.
\code{vertical_list} is the list includes formulas for converting basic horizontal bars to vertical bars.
}
\details{
It does not actually realize the vertical writing system, but actually just changes lines one character at a time.
If horizontal bars are not replaced with vertical bars, unnatural Japanese notation may result.
}
